backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pytesseract==0.3.10
Pillow==10.1.0
spacy==3.7.2
scikit-learn==1.3.2
python-dateutil==2.8.2
pandas==2.1.3
statsmodels==0.14.0
python-multipart==0.0.6
Jinja2==3.1.2
numpy==1.24.3
joblib==1.3.2
apscheduler==3.10.4
email-validator>=2.0.0
slowapi==0.1.9